// src/Components/Interface.ts
export interface Users {
  login: string;
  title: string;
  avatar_url: string;
  id: number;
  name: string;
  company: string;
  blog: string;
  location: string;
  twitter_username: string;
  followers: number;
  following: number;
  created_at: string;
  updated_at: string;
  followers_url: string;
  public_repos: number;
  bio: null;
  url: string;
}